fix: compile sleigh only if not present

+++ build_sleigh.sh
@@ -1,7 +1,9 @@
 #!/bin/bash
-git submodule init
-git submodule update
-cd deps/sleigh
-cmake -B build -S .
-cmake --build build --parallel 8
-cmake --install build --prefix ./../../vendor/
+if [ ! -f "./vendor/lib/libsla.a" ]; then
+    git submodule init
+    git submodule update
+    cd deps/sleigh
+    cmake -B build -S .
+    cmake --build build --parallel 8
+    cmake --install build --prefix ./../../vendor/
+fi
